requirements.txt
fastapi
uvicorn[standard]
pandas
google-cloud-bigquery
db-dtypes